upkeep/requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
google-api-core==2.4.0
google-auth==2.3.3
google-cloud-core==2.2.1
google-cloud-storage==2.0.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.9.0
joblib==1.1.0
lightgbm==3.3.2
numpy==1.21.4
pandas==1.3.5
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
SQLAlchemy==1.4.28
threadpoolctl==3.0.0
typing_extensions==4.0.1
urllib3==1.26.8
zipp==3.6.0
